Create GitHub Actions syntax for one action

Thin wrapper around GitHub actions.

action(IDENTIFIER, needs = NULL, uses, runs = NULL, args = NULL,
  env = NULL, secrets = NULL)

action2hcl(l)

action2docker(l, ...)

Arguments

IDENTIFIER
[character(1)] Giving the name of the action.

Used:

as an informative label on GitHub.com,

in the needs fields of other action blocks to model the workflow graph,

in the resolves fields of other workflow blocks to model the workflow graph.

needs
[character()] giving the actions (by their IDENTIFIERs) that must complete successfully before this action will be invoked. Defaults to NULL for no upstream dependencies.

uses
[character(1)] giving the Docker image that will run the action.

runs
[character()] giving the command to run in the docker image. Overrides the Dockerfile ENTRYPOINT. Defaults to NULL for the default ENTRYPOINT (recommended).

args
[character()] giving the arguments to pass to the action. Arguments get appended to the last command in ENTRYPOINT. Defaults to NULL for no arguments.

env
[list(character(1)] giving the environment variables to set in the action's runtime environment. Defaults to NULL for no environment variables (in addition to the defaults set by GitHub Actions).

secrets
[character()] giving the names of the secret variables to set in the runtime enviornment, which the action can access as an environment variable. The values of secrets must be set in your repository's "Settings" tab. Do not store secrets in your repository. GitHub advises against using GitHub actions for production secrets during the public beta period. Defaults to NULL for no secrets.

l
[list()] giving the named list of HCL fields as returned by action() and workflow().

...
Arguments passed on to processx::run

error_on_status

Whether to throw an error if the command returns with a non-zero status, or it is interrupted. The error clases are system_command_status_error and system_command_timeout_error, respectively, and both errors have class system_command_error as well. See also "Error conditions" below.

wd

Working directory of the process. If NULL, the current working directory is used.

echo_cmd

Whether to print the command to run to the screen.

echo

Whether to print the standard output and error to the screen. Note that the order of the standard output and error lines are not necessarily correct, as standard output is typically buffered.

spinner

Whether to show a reassuring spinner while the process is running.

timeout

Timeout for the process, in seconds, or as a difftime object. If it is not finished before this, it will be killed.

stdout_line_callback

NULL, or a function to call for every line of the standard output. See stdout_callback and also more below.

stdout_callback

NULL, or a function to call for every chunk of the standard output. A chunk can be as small as a single character. At most one of stdout_line_callback and stdout_callback can be non-NULL.

stderr_line_callback

NULL, or a function to call for every line of the standard error. See stderr_callback and also more below.

stderr_callback

NULL, or a function to call for every chunk of the standard error. A chunk can be as small as a single character. At most one of stderr_line_callback and stderr_callback can be non-NULL.

stderr_to_stdout

Whether to redirect the standard error to the standard output. Specifying TRUE here will keep both in the standard output, correctly interleaved. However, it is not possible to deduce where pieces of the output were coming from. If this is TRUE, the standard error callbacks (if any) are never called.

env

Environment of the child process, a named character vector. IF NULL, the environment of the parent is inherited.

windows_verbatim_args

Whether to omit the escaping of the command and the arguments on windows. Ignored on other platforms.

windows_hide_window

Whether to hide the window of the application on windows. Ignored on other platforms.

encoding

The encoding to assume for stdout and stderr. By default the encoding of the current locale is used. Note that processx always reencodes the output of both streams in UTF-8 currently.

cleanup_tree

Whether to clean up the child process tree after the process has finished.

Value

[list()] list of action attributes.

Details

For details on the syntax and arguments, see here

These functions are for advanced users knowledgeable about GitHub actions. Novice users may be better served by the complete templates in workflows.

These functions provide very thin wrappers around existing GitHub actions, including actions from other repositories. Essentially, they just create lists ready to be ingested by action2hcl(), which then turns these R lists into valid GitHub actions syntax blocks.

For documentation on these actions, consult their respective README.mds linked in the below. Some variants of these action wrappers include sensible defaults for frequent uses in R.

The uses field is always hardcoded to a particular commit or tag of the underlying github action to ensure compatibility.

To render an action block completely from scratch, you can always use the templating function action().

Functions

action2hcl: Convert action to HCL

action2docker: Construct corresponding docker run command for an action.

Examples

# many R projects will need this block to first build an image from a DOCKERFILE
l <- action(
  IDENTIFIER = "Add two numbers",
  uses = "rocker/r-ver:3.6.1",
  args = "Rscript -e '1+1'"
)
action2hcl(l = l)

#> action "Add two numbers" {
#> 
#>   uses = "rocker/r-ver:3.6.1"
#>   args = [
#>     "Rscript -e '1+1'"
#>   ]
#> }
#>