Create Google Firebase CLI action to use Firebase

Remember to provide FIREBASE_TOKEN as a secret to the GitHub UI.

Thin wrapper around GitHub actions.

Configuration details other than PROJECT_ID are read from the firebase.json at the root of your repository.

firebase(IDENTIFIER, needs, args, PROJECT_ID = NULL)

firebase_deploy(IDENTIFIER = "Deploy", needs, PROJECT_ID = NULL)

Arguments

IDENTIFIER
[character(1)] giving the name of the action or workflow block.

Used:

as an informative label on GitHub.com,

in the needs fields of other action blocks to model the workflow graph,

in the resolves fields of other workflow blocks to model the workflow graph.

needs
[character()] giving the actions (by their IDENTIFIERs) that must complete successfully before this action will be invoked. Defaults to NULL for no upstream dependencies.

args
[character()] giving the arguments to pass to the action. Arguments get appended to the last command in ENTRYPOINT. Defaults to NULL for no arguments.

PROJECT_ID
[character(1)] giving a specific project to use for all commands, not required if you specify a project in your `.firebaserc`` file.

...
arguments passed on to other methods, not currently used.

Details

These functions are for advanced users knowledgeable about GitHub actions. Novice users may be better served by the complete templates in workflows.

These functions provide very thin wrappers around existing GitHub actions, including actions from other repositories. Essentially, they just create lists ready to be ingested by make_action_block(), which then turns these R lists into valid GitHub actions syntax blocks.

For documentation on these actions, consult their respective README.mds linked in the below. Some variants of these action wrappers include sensible defaults for frequent uses in R.

The uses field is always hardcoded to a particular commit or tag of the underlying github action to ensure compatibility.

To render an action block completely from scratch, you can always use the templating function make_action_block().

Because firebase gets the deploy directory from a firebase.json file, it cannot automatically find the appropriate path. Manually edit your firebase.json to provide the deploy path.

Functions

firebase_deploy: Deploy static assets to Firebase Hosting

See also