Fix documentation

This GitHub action creates man/ documentation from roxygen comments in R/ scripts at the repository root using devtools.

fix_docs(IDENTIFIER = "Fix Documentation", after_code = NULL)

Arguments

IDENTIFIER
[character(1)] giving the name of the workflow block.

after_code
[character(1)] Giving what happens when the working tree is unclean after code is evaluated:

NULL to throw an error or

"commit" to commit the changes. Defaults to NULL, which just thinly wraps check_clean_tree().

Value

[list()] A list as specified in the workflow argument to use_ghactions().

Details

If you set after_code = 'commit' this action will automatically commit and push changes to your repository for you. This will pollute your commit history and may cause unintended interruptions, such as merge conflicts with yourself. The programmatic commit will not trigger another action run, but may trigger other workflow automations (such as Travis and AppVeyor).

GitHub actions are currently available only in repos who belong to organisations or personal accounts who are on the beta. GitHub actions always runs against the repo to which the push was made, and does not currently support pull requests.

For more caveats, see auto_commit().

Examples

workflow(
  IDENTIFIER = "Run calculation",
  on = "push",
  resolves = "Simple Addition"
)

#> $IDENTIFIER
#> [1] "Run calculation"
#> 
#> $on
#> [1] "push"
#> 
#> $resolves
#> [1] "Simple Addition"
#> 
#> $template
#> [1] "workflow"
#>