Create rsync action to deploy via Rsync over SSH

Remember to provide SSH_PRIVATE_KEY and SSH_PUBLIC_KEY as secrets to the GitHub UI..

rsync(IDENTIFIER, needs, HOST_NAME, HOST_IP, HOST_FINGERPRINT, SRC, USER,
  DEST, env = NULL, args = NULL)

Arguments

IDENTIFIER
[character(1)] Giving the name of the action.

Used:

as an informative label on GitHub.com,

in the needs fields of other action blocks to model the workflow graph,

in the resolves fields of other workflow blocks to model the workflow graph.

needs
[character()] giving the actions (by their IDENTIFIERs) that must complete successfully before this action will be invoked. Defaults to NULL for no upstream dependencies.

HOST_NAME
[character(1)] giving the name of the server you wish to deploy to, such as foo.example.com.

HOST_IP
[character(1)] giving the IP of the server you wish to deploy to, such as 111.111.11.111.

HOST_FINGERPRINT
[character(1)] giving the fingerprint of the server you wish to deploy to, can have different formats.

SRC
[character(1)] giving the source directory, relative path from /github/workspace without trailing slash.

USER
[character(1)] giving the user at the target HOST_NAME.

DEST
[character(1)] giving the directory from the root of the HOST_NAME target to write to.

env
[list(character(1)] giving the environment variables to set in the action's runtime environment. Defaults to NULL for no environment variables (in addition to the defaults set by GitHub Actions).

args
[character()] giving the arguments to pass to the action. Arguments get appended to the last command in ENTRYPOINT. Defaults to NULL for no arguments.

Value

[list()] list of action attributes.

Details

For details on the syntax and arguments, see here

These functions are for advanced users knowledgeable about GitHub actions. Novice users may be better served by the complete templates in workflows.

These functions provide very thin wrappers around existing GitHub actions, including actions from other repositories. Essentially, they just create lists ready to be ingested by action2hcl(), which then turns these R lists into valid GitHub actions syntax blocks.

For documentation on these actions, consult their respective README.mds linked in the below. Some variants of these action wrappers include sensible defaults for frequent uses in R.

The uses field is always hardcoded to a particular commit or tag of the underlying github action to ensure compatibility.

To render an action block completely from scratch, you can always use the templating function action().

Examples

# many R projects will need this block to first build an image from a DOCKERFILE
l <- action(
  IDENTIFIER = "Add two numbers",
  uses = "rocker/r-ver:3.6.1",
  args = "Rscript -e '1+1'"
)
action2hcl(l = l)

#> action "Add two numbers" {
#> 
#>   uses = "rocker/r-ver:3.6.1"
#>   args = [
#>     "Rscript -e '1+1'"
#>   ]
#> }
#>