Workflow automation with GitHub Actions

Sets up workflow automation, including continuous integration and deployment (CI/CD) for different kinds of R projects on GitHub actions. This function

Picks a set of sensible defaults for your project.

Transforms a list of workflow and action blocks into the GitHub actions syntax.

Adds a .github/main.workflow file to your repository.

use_ghactions(workflow = website())

list2ghact(workflow)

Arguments

workflow
[list(list())] A named list of blocks nested as:

The workflow block: arguments to workflow() as a named list and

`$actions``, which in turn comprises of the

Several action blocks with arguments to action() as a named list.

Value

A logical vector indicating if file was modified.

Functions

list2ghact: Helper to turn block lists into strings. Useful when you want the result printed to console, not written to file.

See also

Examples

# NOT RUN {
use_ghactions(workflow = website())
# }
# this will print the result to the console for inspection

# NOT RUN {
list2ghact(workflow = website())
# }